Разделена логика редьюсера
