+++ src/components/app/app.jsx
@@ -5,12 +5,13 @@ import GenreQuestionScreen from "../genre-question-screen/genre-question-screen"
 import ArtistQuestionScreen from "../artist-question-screen/artist-question-screen";
 import {artistQuestion, genreQuestion} from "../../common/global-prop-types";
 import {connect} from "react-redux";
-import {ActionCreator} from "../../reducer/reducer";
+import {ActionCreator} from "../../reducer/game/game";
 import GameScreen from "../game-screen/game-screen";
 import FailScreen from "../fail-screen/fail-screen";
 import withActivePlayer from "../../hocs/with-active-player/with-active-player";
 import withUserAnswer from "../../hocs/with-user-answer/with-user-answer";
 import {ScreenSteps} from "../../common/constants";
+import NameSpace from "../../reducer/name-spaces";
 
 const GenreQuestionScreenWrapped = withUserAnswer(withActivePlayer(GenreQuestionScreen));
 const ArtistQuestionScreenWrapped = withActivePlayer(ArtistQuestionScreen);
@@ -84,10 +85,10 @@ App.propTypes = {
 };
 
 const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
-  questions: state.questions,
-  step: state.step,
-  mistakes: state.mistakes,
-  gameTime: state.time,
+  questions: state[NameSpace.DATA].questions,
+  step: state[NameSpace.GAME].step,
+  mistakes: state[NameSpace.GAME].mistakes,
+  gameTime: state[NameSpace.GAME].time,
 });
 
 const mapDispatchToProps = (dispatch) => ({
