+++ src/index.js
@@ -3,12 +3,12 @@ import React from "react";
 import App from "./components/app/app";
 import settings from "./mocks/settings";
 import {applyMiddleware, createStore} from "redux";
-import {reducer} from "./reducer/reducer";
+import reducer from "./reducer/reducer";
 import {Provider} from "react-redux";
 import configureAPI from "./api";
 import {compose} from "recompose";
 import thunk from "redux-thunk";
-import Operation from "./reducer/operation";
+import Operation from "./reducer/operation/operation";
 
 const init = ({errorCount, gameTime}) => {
   const api = configureAPI((...args) => store.dispatch(...args));
